requirements.txt
git+https://github.com/AlfioEmanueleFresta/practical-one-time-pad-c-extension.git
git+https://github.com/AlfioEmanueleFresta/practical-dictionary.git
pycrypto
